feat(core): Move Fourier bootstrap buffers from key to CoreEngine
The buffers being stored into the key made the implementation
use `RefCell` to achieve interior mutability which as a consequence,
made the key non `Send + Sync`.

This commit moves the buffers into the `CoreEngine` So that the key can
become `Send` and `Sync`. And the `CoreEngine` stays `Send` and `Sync`
as it does not need to use have interior mutability.

- Move the fourier bootstrap buffers from the `FourierBootstrapKey`
  into the `CoreEngine`.

- The `CoreEngine` has 2 map which store buffers, one for `u32`
  bootstraps the other for `u64`.

- Remove `SurrogateBsk` as the  `FourierBootstrapKey` can now derive
  `Serialize`, `Deserialize`.

- Remove the `backends::core::private::crypto::bootstrap::Bootstrap`
  trait